Closes #517: Move active test code to more appropriate pkgs.
