+++ app/src/androidTest/java/org/mozilla/focus/helpers/SessionLoadedIdlingResource.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.activity.helpers;
+package org.mozilla.focus.helpers;
 
 import android.support.test.espresso.IdlingResource;
